+++ convey/__init__.py
@@ -1,2 +1,5 @@
 # import Config file as the very first one (setup logging)
 from .config import Config
+from .identifier import Subtype
+
+__all__ = [Subtype]